+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.8.1-dev
+
+* **BREAKING BEHAVIOR**: Don't use `PUB_HOSTED_URL` for package downloads,
+  as it has not worked out in practice. Instead, we've added a `--hosted-url`
+  command line argument.
+
 ## 0.8.0
 
 * `PackageAnalyzer.inspectPackage` added a named argument